+++ narrenschiff/modules/gcloud.py
@@ -1,5 +1,3 @@
-import subprocess
-
 from narrenschiff.modules.common import NarrenschiffModule
 
 
@@ -8,7 +6,7 @@ class Gcloud(NarrenschiffModule):
 
     gcloud_cmd = 'gcloud'
 
-    def execute(self):
+    def get_cmd(self):
         command = self.command.get('command')
         args = self.command.get('args', {})
         switches = self.command.get('opts', [])
@@ -19,7 +17,4 @@ class Gcloud(NarrenschiffModule):
 
         flags.extend(['--{}'.format(switch) for switch in switches])
 
-        cmd = ' '.join([Gcloud.gcloud_cmd, command, *flags])
-
-        output, rc = self.subprocess(cmd)
-        self.echo(output, rc)
+        return ' '.join([Gcloud.gcloud_cmd, command, *flags])