+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Gcloud module will fail with commands that require user input, use `--quiet` in opts make these operations pass (**breaking change**)
 * Modules are refactored, all subprocess execution is implemented in the parent module
+* Execute is now parent module method, and not an abstract method. Instead, `get_cmd` is an abstract method that module needs to override an implement. This method needs to return a command that needs to be executed in a shell
 
 # 2.1.0 - 24.12.2020.